feat(project): provide Base{Viewer|Modeler} distributions
This allows users to use the viewer / modeler features without
inheriting our modules.

Related to #258
