+++ lib/Modeler.js
@@ -1,9 +1,8 @@
 import inherits from 'inherits';
 
-import Ids from 'ids';
+import BaseModeler from './BaseModeler';
 
 import Viewer from './Viewer';
-
 import NavigatedViewer from './NavigatedViewer';
 
 import KeyboardMoveModule from 'diagram-js/lib/navigation/keyboard-move';
@@ -131,21 +130,11 @@ var initialDiagram =
  * @param {Array<didi.Module>} [options.additionalModules] a list of modules to use with the default modules
  */
 export default function Modeler(options) {
-  Viewer.call(this, options);
-
-  // hook ID collection into the modeler
-  this.on('import.parse.complete', function(event) {
-    if (!event.error) {
-      this._collectIds(event.definitions, event.context);
-    }
-  }, this);
-
-  this.on('diagram.destroy', function() {
-    this.get('moddle').ids.clear();
-  }, this);
+  BaseModeler.call(this, options);
 }
 
-inherits(Modeler, Viewer);
+inherits(Modeler, BaseModeler);
+
 
 Modeler.Viewer = Viewer;
 Modeler.NavigatedViewer = NavigatedViewer;
@@ -159,42 +148,6 @@ Modeler.prototype.createDiagram = function(done) {
   return this.importXML(initialDiagram, done);
 };
 
-/**
- * Create a moddle instance, attaching ids to it.
- *
- * @param {Object} options
- */
-Modeler.prototype._createModdle = function(options) {
-  var moddle = Viewer.prototype._createModdle.call(this, options);
-
-  // attach ids to moddle to be able to track
-  // and validated ids in the BPMN 2.0 XML document
-  // tree
-  moddle.ids = new Ids([ 32, 36, 1 ]);
-
-  return moddle;
-};
-
-/**
- * Collect ids processed during parsing of the
- * definitions object.
- *
- * @param {ModdleElement} definitions
- * @param {Context} context
- */
-Modeler.prototype._collectIds = function(definitions, context) {
-
-  var moddle = definitions.$model,
-      ids = moddle.ids,
-      id;
-
-  // remove references from previous import
-  ids.clear();
-
-  for (id in context.elementsById) {
-    ids.claim(id, context.elementsById[id]);
-  }
-};
 
 Modeler.prototype._interactionModules = [
 
@@ -242,6 +195,7 @@ Modeler.prototype._modelingModules = [
 // - modeling modules
 
 Modeler.prototype._modules = [].concat(
-  Modeler.prototype._modules,
+  Viewer.prototype._modules,
   Modeler.prototype._interactionModules,
-  Modeler.prototype._modelingModules);
+  Modeler.prototype._modelingModules
+);
