+++ lib/NavigatedViewer.js
@@ -6,6 +6,7 @@ import KeyboardMoveModule from 'diagram-js/lib/navigation/keyboard-move';
 import MoveCanvasModule from 'diagram-js/lib/navigation/movecanvas';
 import ZoomScrollModule from 'diagram-js/lib/navigation/zoomscroll';
 
+
 /**
  * A viewer that includes mouse navigation facilities
  *
@@ -17,6 +18,7 @@ export default function NavigatedViewer(options) {
 
 inherits(NavigatedViewer, Viewer);
 
+
 NavigatedViewer.prototype._navigationModules = [
   KeyboardMoveModule,
   MoveCanvasModule,
@@ -24,5 +26,6 @@ NavigatedViewer.prototype._navigationModules = [
 ];
 
 NavigatedViewer.prototype._modules = [].concat(
-  NavigatedViewer.prototype._modules,
-  NavigatedViewer.prototype._navigationModules);
+  Viewer.prototype._modules,
+  NavigatedViewer.prototype._navigationModules
+);
